Convert Atom to a GADT to restrict use of 'Variable' (#23)
Replaces Atom with Atom a where a is either Val or Var.
Atom Val is the type of Atoms that are 'concrete', there are no
variables, all the values are statically known.
Atom Var is the type of Atoms that may not be concrete. This includes
variables and predicates with variables in them.

Note: Rules, when added, will be Atom Val (they are concrete) but they
contain a set of Atom Var (they are general, but do not need to be
resolved until they are applied).

Rules may need two forms for quantification: Forall, Exists.